GH-96 -- refactored addSite component, actions, and reducers to now support both Add and Edit forms and events
