+++ src/actions/index.js
@@ -1,6 +1,6 @@
 export * from './AuthActions';
 export * from './MapActions';
-export * from './AddSiteActions';
+export * from './AddEditSiteActions';
 export * from './PermissionsActions';
 export * from './SearchAndFilterActions';
 export * from './FavoritesActions';
