+++ src/reducers/index.js
@@ -2,14 +2,14 @@ import {combineReducers} from 'redux';
 
 import AuthReducer from './AuthReducer';
 import MapReducer from './MapReducer';
-import AddSiteReducer from './AddSiteReducer';
+import AddEditSiteReducer from './AddEditSiteReducer';
 import PermissionsReducer from './PermissionsReducer';
 import FavoritesReducer from './FavoritesReducer';
 
 export default combineReducers({
     auth: AuthReducer,
     map: MapReducer,
-    addSite: AddSiteReducer,
+    addEditSite: AddEditSiteReducer,
     permissions: PermissionsReducer,
     favorites: FavoritesReducer
 });
